telegram_bot_shameless/requirements.txt
aiogram==3.0.0b7
python-dotenv==0.21.1
mysql-connector-python
